Extract types.go to address.go and hash.go
