+++ types/address_test.go
@@ -1,8 +1,6 @@
 package types
 
 import (
-	"math/big"
-	"reflect"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -56,25 +54,5 @@ func TestEIP55(t *testing.T) {
 			addr := StringToAddress(c.address)
 			assert.Equal(t, c.expected, addr.String())
 		})
-	}
-}
-
-func TestTransactionCopy(t *testing.T) {
-	addrTo := StringToAddress("11")
-	txn := &Transaction{
-		Nonce:    0,
-		GasPrice: big.NewInt(11),
-		Gas:      11,
-		To:       &addrTo,
-		Value:    big.NewInt(1),
-		Input:    []byte{1, 2},
-		V:        big.NewInt(25),
-		S:        big.NewInt(26),
-		R:        big.NewInt(27),
-	}
-	newTxn := txn.Copy()
-
-	if !reflect.DeepEqual(txn, newTxn) {
-		t.Fatal("[ERROR] Copied transaction not equal base transaction")
 	}
 }
